+++ src/main/java/lotto/domain/LottoPurchaseAmount.java
@@ -1,46 +1,18 @@
 package lotto.domain;
 
-import java.util.Objects;
 import lotto.view.OutputView;
+
+import java.util.Objects;
 
 public class LottoPurchaseAmount  {
     public static final int LOTTO_PRICE = 1000;
-    private static final int FIRST_CHAR_INDEX = 0;
-    private static final int SECOND_CHAR_INDEX = 1;
-    private static final char NEGATIVE_CHAR = '-';
 
     private final int amount;
 
     public LottoPurchaseAmount(String inputAmount) {
-        validFirstChar(inputAmount);
-        validString(inputAmount);
         this.amount = parseAmount(inputAmount);
-        validPositive();
         validMinAmount();
         validThousands();
-    }
-
-    private void validFirstChar(String inputAmount) {
-        char firstChar = inputAmount.charAt(FIRST_CHAR_INDEX);
-        if (isValidFirstChar(firstChar)) {
-            throw new IllegalArgumentException(OutputView.ERROR_MESSAGE_INPUT_ONLY_NUMBER);
-        }
-    }
-
-    private static boolean isValidFirstChar(char firstChar) {
-        return !(firstChar == NEGATIVE_CHAR || Character.isDigit(firstChar));
-    }
-
-    private void validString(String inputAmount) {
-        for (int i = SECOND_CHAR_INDEX; i < inputAmount.length(); i++) {
-            isDigit(inputAmount, i);
-        }
-    }
-
-    private static void isDigit(String inputAmount, int i) {
-        if (!Character.isDigit(inputAmount.charAt(i))) {
-            throw new IllegalArgumentException(OutputView.ERROR_MESSAGE_INPUT_ONLY_NUMBER);
-        }
     }
 
     private int parseAmount(String inputAmount) {
@@ -54,12 +26,6 @@ public class LottoPurchaseAmount  {
     private void validThousands() {
         if (this.amount % LOTTO_PRICE > 0) {
             throw new IllegalArgumentException(OutputView.ERROR_MESSAGE_AMOUNT_UNIT_OF_1000);
-        }
-    }
-
-    private void validPositive() {
-        if (this.amount <= 0) {
-            throw new IllegalArgumentException(OutputView.ERROR_MESSAGE_MINIMUM_PURCHASE_AMOUNT);
         }
     }
