+++ src/main/java/lotto/domain/WinningNumbers.java
@@ -1,7 +1,8 @@
 package lotto.domain;
 
-import java.util.Objects;
 import lotto.view.OutputView;
+
+import java.util.Objects;
 
 public class WinningNumbers {
     private final LottoNumbers winningNumbers;
@@ -10,19 +11,19 @@ public class WinningNumbers {
     private WinningNumbers(LottoNumbers winningNumbers, LottoNumber bonusBall) {
         this.winningNumbers = winningNumbers;
         this.bonusBall = bonusBall;
+        validDuplicateBonusNumber();
+    }
+
+    private void validDuplicateBonusNumber() {
+        if (this.winningNumbers.isContainNumber(this.bonusBall)) {
+            throw new IllegalArgumentException(OutputView.ERROR_MESSAGE_BONUS_NUMBER_NOT_DUPLICATE_LOTTO_NUMBER);
+        }
     }
 
     public static WinningNumbers of(String input, int inputBonus) {
         LottoNumbers lottoNumbers = LottoNumbers.of(new ManualNumberGenerator(input));
         LottoNumber bonusNumber = new LottoNumber(inputBonus);
-        validDuplicateBonusNumber(lottoNumbers, bonusNumber);
         return new WinningNumbers(lottoNumbers, bonusNumber);
-    }
-
-    private static void validDuplicateBonusNumber(LottoNumbers lottoNumbers, LottoNumber bonusNumber) {
-        if (lottoNumbers.isContainNumber(bonusNumber)) {
-            throw new IllegalArgumentException(OutputView.ERROR_MESSAGE_BONUS_NUMBER_NOT_DUPLICATE_LOTTO_NUMBER);
-        }
     }
 
     public WinningRank matchWinningRank(LottoNumbers lottoNumbers) {
